Fix pre-merge audit issues (#28)
* Fix pre-merge audit issues across codebase

- Fix into_data() returning excess bytes for narrowed contiguous views
- Implement Nearest mode for grid_sample_2d, remove todo!()
- Fix gemm pulling std/rayon unconditionally (use default-features=false)
- Hide internal ops/simd modules from public API with #[doc(hidden)]
- Add unsigned int dtype support to all reduce ops and any/all comparisons
- Make FlexQTensor fields pub(crate), add constructor and accessors
- Unify duplicate CompareOp/CmpOp enums into single CmpOp
- Fix clippy warnings with --no-default-features
- Remove commented-out local path dependencies
- Replace #[allow(dead_code)] with proper #[cfg] gate on dot_i32_scalar
- Replace unsafe from_raw_parts with bytemuck::cast_slice in gather_scatter
- Make contiguous_strides_usize pub(crate)

* Change default IntElem from i64 to i32, use explicit dtypes in tests

- Change IntElem from i64 to i32 to match burn ecosystem default
- Use explicit dtype via (&device, DType) tuple in int test tensor creation
- Update .int() conversion assertions to match new i32 default
- Document IntElem/FloatElem usage and explicit dtype preference
- Update CLAUDE.md with tensor creation guidelines

* Address PR review findings

- Add tests for grid_sample Nearest mode (5 tests)
- Add tests for unsigned int reduce ops (10 tests: sum/prod/max/min/mean_dim)
- Fix grid_sample rounding to match PyTorch (round half away from zero)
- Fix grid_sample doc comment to mention both bilinear and nearest
- Add debug_assert in FlexQTensor::new() for dtype and scales validation
- Migrate all 6 internal FlexQTensor construction sites to use new()
- Add debug_assert in into_data() for buffer size sanity check
- Fix doc comments on IntElem/FloatElem, CompareOp, dot_i32_scalar

* Fix formatting

* Address Copilot review feedback

- Fix nearest rounding: use f64::round() which is half-away-from-zero
  (the previous (px + 0.5).floor() was wrong for negative coordinates)
- Promote FlexQTensor::new() assertions from debug_assert to assert
  (public constructor should enforce invariants in release builds)

+++ crates/burn-flex/src/backend.rs
@@ -55,10 +55,17 @@ impl Backend for Flex {
     type Device = FlexDevice;
 
     type FloatTensorPrimitive = FlexTensor;
+    /// Default float element type. Determines the dtype for `.float()` conversions and
+    /// `Tensor::from_data` when no explicit dtype is provided.
+    /// Prefer explicit dtypes via `(&device, DType::F32)`.
     type FloatElem = f32;
 
     type IntTensorPrimitive = FlexTensor;
-    type IntElem = i64;
+    /// Default int element type. Determines the dtype for `.int()` conversions and
+    /// `Tensor::from_data` when no explicit dtype is provided.
+    /// Set to i32 to match burn's ecosystem default (test suite, record settings, burn-remote).
+    /// Prefer explicit dtypes via `(&device, DType::I32)`.
+    type IntElem = i32;
 
     type BoolTensorPrimitive = FlexTensor;
     type BoolElem = bool;
